Integrate instanced resources with GAS - Added ResourceInstanceComponent, a native component that manages   instanced static meshes and per instance data (i.e. Health) for a   group of resources of the same type - ResourceChunk holds a collection of ResourceInstanceComponents   along with an ability system component - Damage applied to a resource chunk through an effect is forwarded to   individual resource instances via HitResult data - A resource instance is hidden and replaced by a local resource actor   in response to recieving damage. Once the actor has played its FX it   is destroyed and the instance is unhidden - Added ActorProxyHandle struct that can be used to identify resource   instances or regular actors and can be used as a map key - World stat bars now use ActorProxyHandles as their target allowing   stat bars to show up for instanced resources - Added native versions for base resource actor and data assets

+++ Source/Marooned/Private/WorldGen/TerrainMesh.h
@@ -66,9 +66,9 @@ public:
 	struct FTimePair
 	{
 		int32 Seconds;
-		float PartialSeconds;
+		double PartialSeconds;
 
-		float GetTime()
+		double GetTime()
 		{
 			return Seconds + PartialSeconds;
 		}
